requirements.txt
PyYAML==3.10
httplib2==0.8
nltk==2.0.4
nltk-lsm==0.0.1
numpy==1.7.1
oauth2==1.5.211
python-twitter==1.0
requests==1.2.3
simplejson==3.3.0
wsgiref==0.1.2
